requirements.txt
uvicorn
mlflow
minio
fastapi[all]
pandas
python-dotenv
scikit-learn